requirements.txt
# Core Dependencies
python-dotenv==1.0.0

# Web Crawling
scrapy==2.11.0
selenium==4.15.2
beautifulsoup4==4.12.2
requests==2.31.0
lxml==4.9.3

# Kafka
confluent-kafka==2.3.0

# Database
psycopg2-binary==2.9.9
sqlalchemy==2.0.23
alembic==1.13.0

# Data Validation
pydantic==2.5.1

# Data Processing
pandas==2.1.3
numpy==1.26.2
python-dateutil==2.8.2

# Sentiment Analysis (Vietnamese NLP)
textblob==0.17.1
vaderSentiment==3.3.2
# underthesea==6.7.0  # Commented out - requires Rust compiler
# transformers==4.35.2  # Uncomment for PhoBERT
# torch==2.1.1  # Uncomment for PhoBERT

# CLI Tools
click==8.1.7

# Utilities
pyyaml==6.0.1
schedule==1.2.0

# Logging
loguru==0.7.2

# Testing
pytest==7.4.3
pytest-cov==4.1.0
pytest-asyncio==0.21.1

# Code Quality
black==23.11.0
flake8==6.1.0
mypy==1.7.1

# API (Optional - Uncomment if needed)
# fastapi==0.104.1
# uvicorn[standard]==0.24.0
# python-multipart==0.0.6

# Monitoring (Optional)
# prometheus-client==0.19.0
